+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stack Tests
 - SoftwareTimer (Timer Lib) inside 14.ADCSensorExample
 - Test for Timer Lib
+- Linear Queue at 19.QueueImplementation
 
 ### Changed